capstone_project/backend/identity/runtime/requirements.txt
bedrock-agentcore
strands-agents
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi
uvicorn
requests
